Activities Code Cleanup
Cleanup Activites for platform upgrade

+++ src/main/java/com/usda/fmsc/twotrails/activities/MapManagerActivity.java
@@ -51,8 +51,6 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
     private boolean inDetails = false;
     private int notifyAdapter = -1;
 
-
-//    @SuppressWarnings("unchecked")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -95,7 +93,6 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
         visibleMaps = new ArrayList<>();
 
         maps.sort(ArcGisMapLayer.Comparator);
-        //Collections.sort(maps);
 
         visibleMaps.addAll(maps);
 
@@ -204,26 +201,11 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home: {
-                finish();
-                break;
-            }
-            case R.id.mmMenuLogin: {
-//                Intent intent = new Intent(getBaseContext(), ArcGisLoginActivity.class);
-//
-//                UserCredentials credentials = getTtAppCtx().getArcGISTools().getCredentials(MapManagerActivity.this);
-//                final String oldUn = credentials != null ? credentials.getUserName() : StringEx.Empty;
-//
-//                if (!StringEx.isEmpty(oldUn)) {
-//                    intent.putExtra(ArcGisLoginActivity.USERNAME, oldUn);
-//                }
-//
-//                startActivityForResult(intent, Consts.Codes.Activites.ARC_GIS_LOGIN);
-
-                Toast.makeText(MapManagerActivity.this, "ArcGIS Login is currently not supported.", Toast.LENGTH_LONG).show();
-                break;
-            }
+        int itemId = item.getItemId();
+        if (itemId == android.R.id.home) {
+            finish();
+        } else if (itemId == R.id.mmMenuLogin) {
+            Toast.makeText(MapManagerActivity.this, "ArcGIS Login is currently not supported.", Toast.LENGTH_LONG).show();
         }
 
         return super.onOptionsItemSelected(item);
@@ -246,9 +228,6 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
 
     public void btnMmAddOnlineClick(View view) {
         fabSheet.hideSheet();
-//        if (checkCredentials()) {
-//            createMap(null, null, NewArcMapDialog.CreateMode.NEW_ONLINE);
-//        }
 
         Toast.makeText(MapManagerActivity.this, "Adding custom online maps is currently not supported.", Toast.LENGTH_LONG).show();
     }
@@ -257,78 +236,6 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
         fabSheet.hideSheet();
         createMap(null, null, NewArcMapDialog.CreateMode.OFFLINE_FROM_FILE); //addOfflineMap();
     }
-
-//    private boolean checkCredentials() {
-////        if (getTtAppCtx().getArcGISTools().hasValidCredentials(MapManagerActivity.this)) {
-////            return true;
-////        } else {
-////            String message;
-////            boolean updateCredentials = false;
-////
-////            if (getTtAppCtx().getArcGISTools().hasCredentials(MapManagerActivity.this)) {
-////                if (getTtAppCtx().getArcGISTools().areCredentialsOutOfDate(MapManagerActivity.this)) {
-////                    message = "Your credentials are out of date. Would you like to update them now?";
-////                    updateCredentials = true;
-////                } else {
-////                    message = "??";
-////                }
-////            } else {
-////                message = "You need credentials before creating an offline map. Would you like to add them now?";
-////            }
-////
-////            UserCredentials credentials = getTtAppCtx().getArcGISTools().getCredentials(MapManagerActivity.this);
-////            final String oldUn = updateCredentials && credentials != null ? credentials.getUserName() : StringEx.Empty;
-////
-////            new AlertDialog.Builder(this)
-////                    .setMessage(message)
-////                    .setPositiveButton(R.string.str_yes, (dialog, which) -> {
-////                        Intent intent = new Intent(getBaseContext(), ArcGisLoginActivity.class);
-////
-////                        if (!StringEx.isEmpty(oldUn)) {
-////                            intent.putExtra(ArcGisLoginActivity.USERNAME, oldUn);
-////                        }
-////
-////                        startActivityForResult(intent, Consts.Codes.Activites.ARC_GIS_LOGIN);
-////                    })
-////                    .setNegativeButton(R.string.str_no, null)
-////                    .show();
-////        }
-////
-////        return false;
-////    }
-
-
-//    private void addOfflineMap() {
-//        new AlertDialog.Builder(this)
-//                .setMessage("Create Offline map from:")
-//                .setPositiveButton("Existing Map", (dialog, which) -> {
-//                    if (checkCredentials()) {
-//                        SelectMapTypeDialog.newInstance(maps, SelectMapTypeDialog.SelectMapMode.ALL_ARC)
-//                                .setOnMapSelectedListener((mapType, mapId) -> {
-//                                    ArcGisMapLayer layer = getTtAppCtx().getArcGISTools().getMapLayer(mapId);
-//
-//                                    if (!layer.isOnline() && StringEx.isEmpty(layer.getUrl())) {
-//                                        new AlertDialog.Builder(getBaseContext())
-//                                                .setMessage("This offline map was not created from an online resource. " +
-//                                                        "You can only create offline maps from online maps or offline" +
-//                                                        "maps which were created from online resources.")
-//                                                .setPositiveButton(R.string.str_ok, null)
-//                                                .show();
-//                                    } else {
-//                                        createMap(String.format("%s (Offline)", layer.getName()), layer.getUrl(),
-//                                                layer.isOnline() ?
-//                                                        NewArcMapDialog.CreateMode.OFFLINE_FROM_ONLINE_URL :
-//                                                        NewArcMapDialog.CreateMode.OFFLINE_FROM_OFFLINE_URL
-//                                        );
-//                                    }
-//                                })
-//                                .show(getSupportFragmentManager(), SELECT_MAP);
-//                    }
-//                })
-//                .setNegativeButton("File", (dialog, which) -> createMap(null, null, NewArcMapDialog.CreateMode.OFFLINE_FROM_FILE))
-//                .setNeutralButton(R.string.str_cancel, null)
-//                .show();
-//    }
 
     private void createMap(String name, String uri, NewArcMapDialog.CreateMode mode) {
         NewArcMapDialog.newInstance(name, uri, mode)
@@ -385,11 +292,11 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
                 tvName.setText(layer.getName());
 
                 ofmbMenu.setListener(item -> {
-                    switch (item.getItemId()) {
-                        case R.id.ctx_menu_rename: {
-                            final InputDialog id = new InputDialog(MapManagerActivity.this);
+                    int itemId = item.getItemId();
+                    if (itemId == R.id.ctx_menu_rename) {
+                        final InputDialog id = new InputDialog(MapManagerActivity.this);
 
-                            id.setInputText(layer.getName())
+                        id.setInputText(layer.getName())
                                 .setPositiveButton(R.string.str_rename, (dialog, which) -> {
                                     getTtAppCtx().getArcGISTools().updateMapLayer(layer);
                                     layer.setName(id.getText());
@@ -397,21 +304,17 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
                                 })
                                 .setNeutralButton(R.string.str_cancel, null)
                                 .show();
-                            break;
-                        }
-                        case R.id.ctx_menu_delete: {
-                            new AlertDialog.Builder(MapManagerActivity.this)
-                                    .setMessage("Arc you sure you want to delete this map?")
-                                    .setPositiveButton(R.string.str_delete, (dialog, which) -> getTtAppCtx().getArcGISTools().deleteMapLayer(MapManagerActivity.this, layer.getId(), true, new IListener() {
-                                        @Override
-                                        public void onEventTriggered(Object o) {
-                                            removeMap(layer, false);
-                                        }
-                                    }))
-                                    .setNeutralButton(R.string.str_cancel, null)
-                                    .show();
-                            break;
-                        }
+                    } else if (itemId == R.id.ctx_menu_delete) {
+                        new AlertDialog.Builder(MapManagerActivity.this)
+                                .setMessage("Arc you sure you want to delete this map?")
+                                .setPositiveButton(R.string.str_delete, (dialog, which) -> getTtAppCtx().getArcGISTools().deleteMapLayer(MapManagerActivity.this, layer.getId(), true, new IListener() {
+                                    @Override
+                                    public void onEventTriggered(Object o) {
+                                        removeMap(layer, false);
+                                    }
+                                }))
+                                .setNeutralButton(R.string.str_cancel, null)
+                                .show();
                     }
 
                     return false;
@@ -421,7 +324,6 @@ public class MapManagerActivity extends CustomToolbarActivity implements ArcGIST
             }
         }
 
-        @SuppressWarnings("unchecked")
         private void viewMapDetails(View view, ArcGisMapLayer agml) {
             Intent i = new Intent(MapManagerActivity.this, MapDetailsActivity.class);
             i.putExtra(Consts.Codes.Data.MAP_DATA, agml);
